requirements.txt
gradio==5.4.0
gradio_client
python-dotenv